# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /incident-reports:
    get:
      operationId: listIncidentReports
      tags: [crypto]
      summary: List incident reports (encrypted content)
      responses:
        "200":
          description: Report metadata with encrypted content blobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EncryptedReport"
    post:
      operationId: createIncidentReport
      tags: [crypto]
      summary: Create a report draft (encrypted content)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EncryptedReportPut"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EncryptedReport"
  /incident-reports/{id}:
    get:
      operationId: getIncidentReport
      tags: [crypto]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Report metadata with encrypted content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EncryptedReport"
    put:
      operationId: updateIncidentReport
      tags: [crypto]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EncryptedReportPut"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EncryptedReport"
  /incident-reports/{id}/submit:
    post:
      operationId: submitIncidentReport
      tags: [crypto]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EncryptedReport"
  /incident-reports/{id}/addendum:
    post:
      operationId: addReportAddendum
      tags: [crypto]
      summary: Append an addendum (client re-encrypts the whole content)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EncryptedContentPut"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EncryptedReport"

  # --- Datenexport: verschluesseltes Buendel, Loeschung nach Bestaetigung ---
...
components:
  schemas:
...
    EncryptedContentPut:
      type: object
      properties:
        contentEncrypted:
          type: string
        contentKeyVersion:
          type: integer
      required:
        - contentEncrypted
        - contentKeyVersion
    EncryptedReport:
      type: object
      properties:
        id:
          type: string
        schoolId:
          type: string
        missionId:
          type: string
          nullable: true
        missionTitle:
          type: string
          nullable: true
        authorId:
          type: string
        status:
          type: string
        location:
          type: string
          nullable: true
        incidentAt:
          type: string
          format: date-time
          nullable: true
        responderIds:
          type: array
          items:
            type: string
          nullable: true
        contentEncrypted:
          type: string
          nullable: true
        contentKeyVersion:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        submittedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - authorId
        - status
        - createdAt
        - updatedAt
    EncryptedReportPut:
      type: object
      properties:
        missionId:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        incidentAt:
          type: string
          format: date-time
          nullable: true
        responders:
          type: array
          items:
            type: string
          nullable: true
        contentEncrypted:
          type: string
        contentKeyVersion:
          type: integer
      required:
        - contentEncrypted
        - contentKeyVersion
...